error fix

+++ data-base/mapper/mapper.go
@@ -5,17 +5,17 @@ import (
 	"time"
 )
 
-func MapLocationToSector(location *structs.Location) int64 {
+func MapLocationToSector(location *structs.Location) int32 {
 
 	return 10
 }
 
-func CalculateDeliveryTime(distance int64) time.Time {
+func CalculateDeliveryTime(distance int32) time.Time {
 
 	return time.Now().UTC()
 }
 
-func CalculateDeliveryCost(distance int64, address *structs.Address) float64 {
+func CalculateDeliveryCost(distance int32, address *structs.Address) float32 {
 
-	return float64(distance) * 5
+	return float32(distance) * 5
 }
